// src/lib/providers/types.ts
export interface SkipInterval {
  title?: string;
  from: number;
  to: number;
}
// ...
export interface Link {
  title: string;
  quality?: string;
  episodesLink?: string;
  quickDownload?: boolean;
  directLinks?: {
    title: string;
    link: string;
    type?: 'movie' | 'series';
    description?: string;
    image?: string;
    quickDownload?: boolean;
    skip?: SkipInterval[];
    /** Filler episode (not part of the main story). */
    filler?: boolean;
  }[];
}